app/requirements.txt
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.31.0
bcrypt==3.2.2
cffi==2.0.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.122.0
h11==0.16.0
idna==3.11
iso8601==2.1.0
passlib==1.7.4
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypika-tortoise==0.6.3
python-dotenv==1.2.1
pytz==2025.2
sniffio==1.3.1
starlette==0.50.0
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
